Update README.md
It's a reminder application for people who has maladies and forget the time of their medicine. The application notify them by the notification to remind them the time of medicine that the user choose it. It's a great solution for this category of people because many times left forgot their medicament  many people died and was ill for this issue and statistiques witness it.